requirements.txt
discord.py
aiofiles
python-dotenv
